Refactor type annotations to use new types module
