--- awaits/typing.py
@@ -1,6 +1,0 @@
-try:
-    from typing import ParamSpec
-except ImportError:
-    from typing_extensions import ParamSpec
-
-P = ParamSpec('P')
